+++ shop/src/main/java/com/phucx/shop/model/CurrentProductList.java
@@ -1,5 +1,7 @@
 package com.phucx.shop.model;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.annotation.Immutable;
 
 import jakarta.persistence.Entity;
@@ -19,7 +21,7 @@ public class CurrentProductList {
     private Integer productID;
     private String productName;
     private String picture;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     private Integer unitsInStock;
     private String discountID;
     private Integer discountPercent;
