+++ shop/src/main/java/com/phucx/shop/model/OrderItem.java
@@ -1,5 +1,6 @@
 package com.phucx.shop.model;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -15,12 +16,14 @@ public class OrderItem {
     private Integer unitsInStock;
     private List<OrderItemDiscount> discounts;
     private String picture;
-    private Double unitPrice;
-    private Double extendedPrice;
+   
+    private BigDecimal unitPrice;
+
+    private BigDecimal extendedPrice;
 
 
 
-    public OrderItem(Integer productID, String productName, String categoryName, Integer quantity, Integer unitsInStock, String picture, Double unitPrice) {
+    public OrderItem(Integer productID, String productName, String categoryName, Integer quantity, Integer unitsInStock, String picture, BigDecimal unitPrice) {
         this();
         this.productName=productName;
         this.productID = productID;
@@ -34,7 +37,7 @@ public class OrderItem {
 
 
     public OrderItem(Integer productID, String productName, String categoryName, Integer quantity, Integer unitsInStock, List<OrderItemDiscount> discounts, String picture,
-            Double unitPrice, Double extendedPrice) {
+            BigDecimal unitPrice, BigDecimal extendedPrice) {
         this.productName = productName;
         this.productID = productID;
         this.categoryName = categoryName;
@@ -50,6 +53,6 @@ public class OrderItem {
 
     public OrderItem() {
         this.discounts = new ArrayList<>();
-        this.extendedPrice = Double.valueOf(0);
+        this.extendedPrice = BigDecimal.valueOf(0);
     }
 }
